Spring stability fixes (#5)
Improved numerical stability for underdamped springs as d->1. Underdamped f->0 has similar issues which this commit does not address, but the fix for that will be a minor extension of this change.
The overdamped case also has issues as d->1 and f->0. I chose not to address that for now because the fix will be more involved.
Also added comments.